Rely on scrollIntoView() when doing feature detection
The Chromium community is moving to standard behavior from the CSSOM
specification [1]. In order to launch the feature and analyze potential
broken URLs, they are counting pages setting the scrollLeft of a RTL
scroller to a positive value. Websites using the current version of
jquery.rtl-scroll will trigger the counter [2]. In order to workaround
that issue, rely on Element.scrollIntoView() instead.

This patch also replaces text with divs and removes scrollbar from
the scroller, in order to make measurements more reliable.

[1] https://www.chromestatus.com/feature/5759578031521792
[2] https://github.com/othree/jquery.rtl-scroll-type/issues/6